cardEntity
