+++ client/package.json
@@ -22,6 +22,7 @@
     "react-router-dom": "^7.2.0",
     "react-slick": "^0.30.3",
     "slick-carousel": "^1.8.1",
+    "sweetalert2": "^11.17.2",
     "tailwindcss": "^4.0.7"
   },
   "devDependencies": {
