--- client/src/components/Home.jsx
@@ -1,13 +1,0 @@
-import Navbar from "./Navbar.jsx"
-
-const Home = () => {
-  return (
-    <>
-     
-      <Navbar/>
-   
-    </>
-  )
-}
-
-export default Home
